Moved CreateMap into utils and made a small update to the build method
